pyproject.toml
[tool.poetry]
name = "gke-service-sample"
version = "1.0.0"
description = "A sample fastapi project for gke"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "gke_service_sample"}]

[tool.poetry.dependencies]
python = ">=3.9"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
schedule = "^1.1.0"
lxml = "^4.9.1"
pandas = "^1.5.3"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^22.10.0"
pylint = "^2.15.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
